+++ crates/rolldown_oxc_utils/src/lib.rs
@@ -4,10 +4,12 @@ mod dummy;
 mod ext;
 mod from_in;
 mod into_in;
+mod oxc_ast;
 mod take_in;
 
 pub use crate::{
-  ast_snippet::AstSnippet, dummy::Dummy, from_in::FromIn, into_in::IntoIn, take_in::TakeIn,
+  ast_snippet::AstSnippet, dummy::Dummy, from_in::FromIn, into_in::IntoIn, oxc_ast::OxcAst,
+  take_in::TakeIn,
 };
-pub use compiler::{OxcCompiler, OxcProgram};
+pub use compiler::OxcCompiler;
 pub use ext::{BindingIdentifierExt, BindingPatternExt, ExpressionExt, StatementExt};
